Modal delet

+++ curso-Jdev/curso-jsp/WebContent/cadastroProduto.jsp
@@ -140,10 +140,10 @@
 								<fmt:setLocale value="pt-BR" />		
 								<td class=""><fmt:formatNumber type="currency" maxFractionDigits="2" currencySymbol="" value="${prod.quantidade}" />	</td> 		
 							 	<td class=""><fmt:formatNumber type="currency" maxFractionDigits="2" currencySymbol="R$" value="${prod.valor}" />	</td>
-								<td><a href="salvarProduto?acao=delete&id=${prod.id}"><button
+								<td><a href="salvarProduto?acao=delete&id=${prod.id}" data-confirm='Tem certeza de que deseja excluir o item selecionado?'><button
 											class="btn btn-danger" title="Excluir">
 											<i class="far fa-trash-alt"></i>
-										</button></a> <a href="salvarProduto?acao=editar&id=${prod.id}"
+										</button></a> <a href="salvarProduto?acao=editar&id=${prod.id}"  
 									class="btn btn-primary" title="Editar"> <i
 										class="fas fa-edit"></i>
 								</a></td>
@@ -198,6 +198,20 @@
 		});
 	    
 	}); 
+	
+	$(document).ready(function() {
+		  $("a[data-confirm]").click(function(ev) {
+		    var href = $(this).attr("href");
+		    if (!$("#confirm-delete").length) {
+		      $("body").append(
+		        '<div class="modal fade" id="confirm-delete" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true"><div class="modal-dialog"><div class="modal-content"><div class="modal-header bg-danger text-white">EXCLUIR PRODUTO <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button></div><div class="modal-body">Tem certeza de que deseja excluir o item selecionado?</div><div class="modal-footer"><button type="button" class="btn btn-success" data-dismiss="modal">Cancelar</button><a class="btn btn-danger text-white" id="dataComfirmOK">Apagar</a></div></div></div></div>'
+		      );
+		    }
+		    $("#dataComfirmOK").attr("href", href);
+		    $("#confirm-delete").modal({ show: true });
+		    return false;
+		  });
+		});
 	
 
 </script>
