+++ curso-Jdev/curso-jsp/WebContent/telefones.jsp
@@ -133,7 +133,7 @@
 								<td><c:out value="${fone.telefone }"></c:out></td>
 								<td><c:out value="${fone.tipo }"></c:out></td>
 
-								<td><a href="salvarTelefones?acao=delete&userId=${userId}&id=${fone.id}"><button
+								<td><a href="salvarTelefones?acao=delete&userId=${userId}&id=${fone.id}" data-confirm='Tem certeza de que deseja excluir o usuário selecionado?'><button
 											class="btn btn-danger" title="Excluir">
 											<i class="far fa-trash-alt"></i>
 										</button></a></td>
@@ -173,5 +173,20 @@
 	
 		return true;
 	}
+	
+	$(document).ready(function() {
+		  $("a[data-confirm]").click(function(ev) {
+		    var href = $(this).attr("href");
+		    if (!$("#confirm-delete").length) {
+		      $("body").append(
+		        '<div class="modal fade" id="confirm-delete" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true"><div class="modal-dialog"><div class="modal-content"><div class="modal-header bg-danger text-white">EXCLUIR TELEFONE <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button></div><div class="modal-body">Tem certeza de que deseja excluir o telefone selecionado?</div><div class="modal-footer"><button type="button" class="btn btn-success" data-dismiss="modal">Cancelar</button><a class="btn btn-danger text-white" id="dataComfirmOK">Apagar</a></div></div></div></div>'
+		      );
+		    }
+		    $("#dataComfirmOK").attr("href", href);
+		    $("#confirm-delete").modal({ show: true });
+		    return false;
+		  });
+		});
+	
 </script>
 </html>
